requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.3.1
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
Brotli==1.1.0
CacheControl==0.14.1
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==43.0.3
cssselect2==0.7.0
defusedxml==0.8.0rc2
Django==5.1.1
django-celery-beat==2.7.0
django-cors-headers==4.4.0
django-extensions==3.2.3
django-filter==24.3
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-gis==1.1
djangorestframework-simplejwt==5.3.1
fcm-django==2.2.1
firebase-admin==6.6.0
fonttools==4.55.8
frozenlist==1.5.0
git-filter-repo==2.47.0
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
kombu==5.4.2
msgpack==1.1.0
multidict==6.1.0
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.28.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyparsing==3.2.0
pyphen==0.17.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
swapper==1.4.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
weasyprint==64.0
webencodings==0.5.1
whitenoise==6.8.2
yarl==1.18.3
zopfli==0.2.3.post1
